server/requirements.txt
bcrypt==4.0.0
boto3==1.24.31
botocore==1.27.31
click==8.1.3
Flask==2.1.3
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
Pillow==9.2.0
PyJWT==2.5.0
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytimeparse==1.1.8
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.10
Werkzeug==2.1.2
